Refactored ClanCommands classes
